requirements.txt
pymongo
yfinance
psycopg2-binary
streamlit
pickle
arch
statsmodels